Enhance mock factories by introducing a sophisticated DI container mock with registration capabilities. Update existing mock configuration to support optional overrides. Refactor tests to utilize the new container mock, improving flexibility and maintainability.
