+++ tests/unit/loaders/baseManifestItemLoader.validatePrimarySchema.test.js
@@ -4,24 +4,11 @@ import { describe, it, expect, jest, beforeEach } from '@jest/globals';
 // Adjust the import path as necessary
 import { BaseManifestItemLoader } from '../../../src/loaders/baseManifestItemLoader.js';
 import { formatAjvErrors } from '../../../src/utils/ajvUtils.js';
-import { createMockPathResolver, createMockDataFetcher } from '../../common/mockFactories/index.js';
+import { createMockConfiguration, createMockPathResolver, createMockDataFetcher } from '../../common/mockFactories/index.js';
 // Assume ValidationResult type is available or mock it if needed for type checking in tests
 // import { ValidationResult } from '../../../src/interfaces/validation.js'; // Example import
 
 // --- Mock Service Factories (Keep as provided, ensure ISchemaValidator has isSchemaLoaded) ---
-
-const createMockConfiguration = (overrides = {}) => ({
-  getModsBasePath: jest.fn().mockReturnValue('./data/mods'),
-  getContentTypeSchemaId: jest.fn(() => 'testSchema'),
-  getSchemaBasePath: jest.fn().mockReturnValue('schemas'),
-  getSchemaFiles: jest.fn().mockReturnValue([]),
-  getWorldBasePath: jest.fn().mockReturnValue('worlds'),
-  getBaseDataPath: jest.fn().mockReturnValue('./data'),
-  getGameConfigFilename: jest.fn().mockReturnValue('game.json'),
-  getModManifestFilename: jest.fn().mockReturnValue('mod-manifest.json'),
-  getContentBasePath: jest.fn((type) => `./data/${type}`),
-  ...overrides,
-});
 
 const createMockSchemaValidator = (overrides = {}) => ({
   validate: jest.fn().mockReturnValue({ isValid: true, errors: null }),
